# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"CreateTicketRequest": {
				"properties": {
					"title": {
						"type": "string",
						"description": "1-200 chars"
					},
					"message": {
						"type": "string",
						"description": "1-5000 chars"
					}
				},
				"required": [
					"title",
					"message"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TicketListItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"last_message_at": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"title",
					"status",
					"created_at",
					"last_message_at"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PaginationMeta": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"total",
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListResponse_TicketListItem_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/TicketListItem"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TicketDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"messages": {
						"items": {
							"properties": {
								"created_at": {
									"type": "string"
								},
								"body": {
									"type": "string"
								},
								"sender": {
									"type": "string"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"created_at",
								"body",
								"sender",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"status",
					"created_at",
					"messages"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ErrorResponse": {
				"properties": {
					"error": {
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"message",
							"code"
						],
						"type": "object"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AddOnSlug": {
				"type": "string",
				"enum": [
					"oven",
					"fridge",
					"windows",
					"chairs"
				]
			},
...
			"FrequencyOption": {
				"type": "string",
				"enum": [
					"one-off",
					"2-3-times-a-week",
					"every-week",
					"every-2-weeks"
				]
			},
...
			"CleaningProducts": {
				"type": "string",
				"enum": [
					"bring",
					"provide"
				]
			},
			"DayOfWeek": {
				"type": "string",
				"enum": [
					"monday",
					"tuesday",
					"wednesday",
					"thursday",
					"friday",
					"saturday",
					"sunday"
				]
			},
			"QuoteLineItem": {
				"properties": {
					"label": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"label",
					"amount"
				],
				"type": "object",
				"additionalProperties": true
			},
			"QuoteResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"postcode": {
						"type": "string"
					},
					"rooms": {
						"type": "number",
						"format": "double"
					},
					"bathrooms": {
						"type": "number",
						"format": "double"
					},
					"add_ons": {
						"items": {
							"$ref": "#/components/schemas/AddOnSlug"
						},
						"type": "array"
					},
					"hours": {
						"type": "number",
						"format": "double"
					},
					"cleaning_products": {
						"$ref": "#/components/schemas/CleaningProducts"
					},
					"frequency": {
						"$ref": "#/components/schemas/FrequencyOption"
					},
					"preferred_days": {
						"items": {
							"$ref": "#/components/schemas/DayOfWeek"
						},
						"type": "array"
					},
					"line_items": {
						"items": {
							"$ref": "#/components/schemas/QuoteLineItem"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string",
						"enum": [
							"GBP"
						],
						"nullable": false
					},
					"recommended_hours": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"postcode",
					"rooms",
					"bathrooms",
					"add_ons",
					"hours",
					"cleaning_products",
					"frequency",
					"preferred_days",
					"line_items",
					"total",
					"currency",
					"recommended_hours"
				],
				"type": "object",
				"additionalProperties": true
			},
			"QuoteRequest": {
				"properties": {
					"postcode": {
						"type": "string",
						"description": "Valid UK postcode e.g. DE1 1AA"
					},
					"rooms": {
						"type": "number",
						"format": "double",
						"description": "Number of rooms to clean (1–20)"
					},
					"bathrooms": {
						"type": "number",
						"format": "double",
						"description": "Number of bathrooms (0–10)"
					},
					"add_ons": {
						"items": {
							"$ref": "#/components/schemas/AddOnSlug"
						},
						"type": "array",
						"description": "Selected add-on services"
					},
					"hours": {
						"type": "number",
						"format": "double",
						"description": "Duration in hours. Valid values: 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0, 7.5, 8.0, 8.5, 9.0"
					},
					"cleaning_products": {
						"$ref": "#/components/schemas/CleaningProducts",
						"description": "'bring' = Fada provides sprays and cloths (not mops, buckets, or vacuums). 'provide' = customer provides all products."
					},
					"frequency": {
						"$ref": "#/components/schemas/FrequencyOption"
					},
					"preferred_days": {
						"items": {
							"$ref": "#/components/schemas/DayOfWeek"
						},
						"type": "array"
					},
					"email": {
						"type": "string"
					},
					"newsletter_opt_in": {
						"type": "boolean"
					}
				},
				"required": [
					"postcode",
					"rooms",
					"bathrooms",
					"hours",
					"cleaning_products",
					"frequency",
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"QuoteScheduleResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"preferred_date": {
						"type": "string"
					},
					"preferred_time": {
						"type": "string"
					},
					"weekend_surcharge": {
						"type": "number",
						"format": "double",
						"description": "Additional charge of £10 applied when booking falls on Friday, Saturday, or Sunday"
					},
					"final_total": {
						"type": "number",
						"format": "double"
					},
					"currency": {
						"type": "string",
						"enum": [
							"GBP"
						],
						"nullable": false
					}
				},
				"required": [
					"id",
					"preferred_date",
					"preferred_time",
					"weekend_surcharge",
					"final_total",
					"currency"
				],
				"type": "object",
				"additionalProperties": true
			},
			"QuoteScheduleRequest": {
				"properties": {
					"preferred_date": {
						"type": "string",
						"description": "ISO date string (YYYY-MM-DD). Must be a future date."
					},
					"preferred_time": {
						"type": "string",
						"description": "Arrival time slot. 30-min intervals between 09:00 and 17:00 e.g. \"10:00\""
					}
				},
				"required": [
					"preferred_date",
					"preferred_time"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"NotificationResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type",
					"title",
					"body",
					"read",
					"created_at"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListResponse_NotificationResponse_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/NotificationResponse"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"Service": {
				"properties": {
					"id": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"image_url": {
						"type": "string",
						"nullable": true
					},
					"icon_url": {
						"type": "string",
						"nullable": true
					},
					"display_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"created_at",
					"updated_at",
					"title",
					"slug",
					"description",
					"image_url",
					"icon_url",
					"display_order"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListResponse_Service_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Service"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"BlogPost": {
				"properties": {
					"id": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"excerpt": {
						"type": "string",
						"nullable": true
					},
					"content": {
						"type": "string"
					},
					"cover_image_url": {
						"type": "string",
						"nullable": true
					},
					"author": {
						"type": "string",
						"nullable": true
					},
					"published_at": {
						"type": "string",
						"nullable": true
					},
					"category_name": {
						"type": "string",
						"nullable": true
					},
					"category_slug": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"created_at",
					"title",
					"slug",
					"excerpt",
					"content",
					"cover_image_url",
					"author",
					"published_at",
					"category_name",
					"category_slug"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListResponse_BlogPost_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/BlogPost"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Faq": {
				"properties": {
					"id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					},
					"display_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"question",
					"answer",
					"display_order"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ListResponse_Faq_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Faq"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CostGuidePricingRow": {
				"properties": {
					"id": {
						"type": "string"
					},
					"service_category": {
						"type": "string"
					},
					"rate": {
						"type": "string"
					},
					"explanation": {
						"type": "string",
						"nullable": true
					},
					"display_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"service_category",
					"rate",
					"explanation",
					"display_order"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CostGuideDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"excerpt": {
						"type": "string",
						"nullable": true
					},
					"published_at": {
						"type": "string",
						"nullable": true
					},
					"content": {
						"type": "string"
					},
					"pricing_table": {
						"items": {
							"$ref": "#/components/schemas/CostGuidePricingRow"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"slug",
					"excerpt",
					"published_at",
					"content",
					"pricing_table"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "fada-lustre-api",
		"version": "1.0.0",
		"description": "Backend API for Fada Lustre — a professional cleaning marketplace based in Derby, UK.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Fada Lustre",
			"email": "marketing@fadalustre.com"
		}
	},
	"paths": {
		"/v1/support/tickets": {
			"post": {
				"operationId": "CreateTicket",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"created_at": {
											"type": "string"
										},
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"created_at",
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new support ticket with a title and initial message.\nAvailable to both customers and cleaners.",
				"summary": "Create support ticket",
				"tags": [
					"Support"
				],
				"security": [
					{
						"jwt": [
							"customer",
							"cleaner"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTicketRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListTickets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_TicketListItem_"
								}
							}
						}
					}
				},
				"description": "List the user's support tickets with optional status filter and pagination.",
				"summary": "List my tickets",
				"tags": [
					"Support"
				],
				"security": [
					{
						"jwt": [
							"customer",
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/support/tickets/{id}": {
			"get": {
				"operationId": "GetTicket",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TicketDetail"
								}
							}
						}
					},
					"404": {
						"description": "Ticket not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve a support ticket with its full message thread.\nOnly accessible by the ticket owner.",
				"summary": "Get ticket details",
				"tags": [
					"Support"
				],
				"security": [
					{
						"jwt": [
							"customer",
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/quotes": {
			"post": {
				"operationId": "CreateQuote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuoteResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Calculate an instant cleaning quote based on postcode, rooms, hours,\ncleaning products preference, and frequency. Returns itemised line items and total.",
				"summary": "Create quote",
				"tags": [
					"Quotes & Pricing"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/QuoteRequest"
							}
						}
					}
				}
			}
		},
		"/v1/quotes/{id}/schedule": {
			"patch": {
				"operationId": "ScheduleQuote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuoteScheduleResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Quote not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Schedule a previously created quote by setting a preferred date and time slot.\nConverts the quote into a bookable appointment.",
				"summary": "Schedule quote",
				"tags": [
					"Quotes & Pricing"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/QuoteScheduleRequest"
							}
						}
					}
				}
			}
		},
...
		"/v1/notifications": {
			"get": {
				"operationId": "ListNotifications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_NotificationResponse_"
								}
							}
						}
					}
				},
				"description": "List notifications for the authenticated user with pagination.\nReturns both read and unread notifications.",
				"summary": "List notifications",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/notifications/{id}/read": {
			"patch": {
				"operationId": "MarkRead",
				"responses": {
					"204": {
						"description": "Marked as read"
					},
					"404": {
						"description": "Notification not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Mark a specific notification as read. Returns 204 on success.",
				"summary": "Mark notification read",
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/services": {
			"get": {
				"operationId": "ListServices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_Service_"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve all active cleaning services. Supports optional text search.\nPublic endpoint — no authentication required.",
				"summary": "List services",
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/services/{slug}": {
			"get": {
				"operationId": "GetService",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Service"
								}
							}
						}
					},
					"404": {
						"description": "Service not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve a single service by its URL-friendly slug.",
				"summary": "Get service by slug",
				"tags": [
					"Services"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/blog/posts": {
			"get": {
				"operationId": "ListPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_BlogPost_"
								}
							}
						}
					}
				},
				"description": "Retrieve published blog posts with optional filtering by category, search,\nand pagination.",
				"summary": "List blog posts",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/blog/posts/{slug}": {
			"get": {
				"operationId": "GetPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogPost"
								}
							}
						}
					},
					"404": {
						"description": "Post not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve a single blog post by its slug, including full content.",
				"summary": "Get blog post",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/faqs": {
			"get": {
				"operationId": "ListFaqs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_Faq_"
								}
							}
						}
					}
				},
				"description": "Retrieve all frequently asked questions. Supports optional text search.",
				"summary": "List FAQs",
				"tags": [
					"FAQs"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/cost-guides/{slug}": {
			"get": {
				"operationId": "GetCostGuide",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CostGuideDetail"
								}
							}
						}
					},
					"404": {
						"description": "Cost guide not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve a single cost guide by slug with full content blocks.",
				"summary": "Get cost guide",
				"tags": [
					"Cost Guides"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "slug",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}